requirements.txt
# Data Handling and Analysis
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.28

# Technical Analysis
ta>=0.11.0
pandas-ta>=0.3.14b0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
imbalanced-learn>=0.11.0

# Visualization
matplotlib>=3.7.0
plotly>=5.17.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
joblib>=1.3.0
tqdm>=4.66.0
